SRV_Channel: more changes for 32 bit servo mask

+++ libraries/SRV_Channel/SRV_Channel.cpp
@@ -80,7 +80,7 @@ SRV_Channel::SRV_Channel(void)
 {
     AP_Param::setup_object_defaults(this, var_info);
     // start with all pwm at zero
-    have_pwm_mask = ~uint16_t(0);
+    have_pwm_mask = ~uint32_t(0);
 }
 
 // convert a 0..range_max to a pwm
